feat: add solutions to lc problem: No.3073
No.3073.Maximum Increasing Triplet Value
