+++ solution/3000-3099/3073.Maximum Increasing Triplet Value/Solution.ts
@@ -1,156 +1,3 @@
-# [3073. 最大递增三元组](https://leetcode.cn/problems/maximum-increasing-triplet-value)
-
-[English Version](/solution/3000-3099/3073.Maximum%20Increasing%20Triplet%20Value/README_EN.md)
-
-<!-- tags: -->
-
-## 题目描述
-
-<!-- 这里写题目描述 -->
-
-<p>给定一个数组&nbsp;<code>nums</code>，返回满足 <code>i &lt; j &lt; k</code> 且 <code>nums[i] &lt; nums[j] &lt; nums[k]</code>&nbsp;的三元组 <code>(i, j, k)</code> <em>&nbsp;</em>的 <strong>最大值</strong><i>。</i></p>
-
-<p>三元组&nbsp;<code>(i, j, k)</code> &nbsp;的&nbsp;<strong>值&nbsp;</strong>为&nbsp;<code>nums[i] - nums[j] + nums[k]</code>。</p>
-
-<div id="gtx-trans" style="position: absolute; left: 274px; top: 102px;">
-<div class="gtx-trans-icon">&nbsp;</div>
-</div>
-
-<p>&nbsp;</p>
-
-<p><strong class="example">示例 1: </strong></p>
-
-<pre>
-输入：nums = [5,6,9]
-输出：8
-解释：对于一个递增的三元组，我们只有一个选择，那就是选择所有三个元素。三元组的值为 5 - 6 + 9 = 8。
-</pre>
-
-<p><strong class="example">示例 2: </strong></p>
-
-<pre>
-输入：nums = [1,5,3,6]
-输出：4
-解释：只有两个递增三元组：
-(0, 1, 3)：这个三元组的值为 nums[0] - nums[1] + nums[3] = 1 - 5 + 6 = 2。
-(0, 2, 3)：这个三元组的值为 nums[0] - nums[2] + nums[3] = 1 - 3 + 6 = 4。
-因此答案是 4。
-</pre>
-
-<p>&nbsp;</p>
-
-<p><strong>提示:</strong></p>
-
-<ul>
-	<li><code>3 &lt;= nums.length &lt;= 10<sup>5</sup></code></li>
-	<li><code>1 &lt;= nums[i] &lt;= 10<sup>9</sup></code></li>
-	<li>输入数据保证至少一个三元组满足给定条件。</li>
-</ul>
-
-## 解法
-
-### 方法一
-
-<!-- tabs:start -->
-
-```python
-from sortedcontainers import SortedList
-
-
-class Solution:
-    def maximumTripletValue(self, nums: List[int]) -> int:
-        n = len(nums)
-        right = [nums[-1]] * n
-        for i in range(n - 2, -1, -1):
-            right[i] = max(nums[i], right[i + 1])
-        sl = SortedList([nums[0]])
-        ans = 0
-        for j in range(1, n - 1):
-            if right[j + 1] > nums[j]:
-                i = sl.bisect_left(nums[j]) - 1
-                if i >= 0:
-                    ans = max(ans, sl[i] - nums[j] + right[j + 1])
-            sl.add(nums[j])
-        return ans
-```
-
-```java
-class Solution {
-    public int maximumTripletValue(int[] nums) {
-        int n = nums.length;
-        int[] right = new int[n];
-        right[n - 1] = nums[n - 1];
-        for (int i = n - 2; i >= 0; --i) {
-            right[i] = Math.max(nums[i], right[i + 1]);
-        }
-        TreeSet<Integer> ts = new TreeSet<>();
-        ts.add(nums[0]);
-        int ans = 0;
-        for (int j = 1; j < n - 1; ++j) {
-            if (right[j + 1] > nums[j]) {
-                Integer it = ts.lower(nums[j]);
-                if (it != null) {
-                    ans = Math.max(ans, it - nums[j] + right[j + 1]);
-                }
-            }
-            ts.add(nums[j]);
-        }
-        return ans;
-    }
-}
-```
-
-```cpp
-class Solution {
-public:
-    int maximumTripletValue(vector<int>& nums) {
-        int n = nums.size();
-        vector<int> right(n, nums.back());
-        for (int i = n - 2; ~i; --i) {
-            right[i] = max(nums[i], right[i + 1]);
-        }
-        set<int> ts;
-        ts.insert(nums[0]);
-        int ans = 0;
-        for (int j = 1; j < n - 1; ++j) {
-            if (right[j + 1] > nums[j]) {
-                auto it = ts.lower_bound(nums[j]);
-                if (it != ts.begin()) {
-                    --it;
-                    ans = max(ans, *it - nums[j] + right[j + 1]);
-                }
-            }
-            ts.insert(nums[j]);
-        }
-        return ans;
-    }
-};
-```
-
-```go
-func maximumTripletValue(nums []int) (ans int) {
-	n := len(nums)
-	right := make([]int, n)
-	right[n-1] = nums[n-1]
-	for i := n - 2; i >= 0; i-- {
-		right[i] = max(nums[i], right[i+1])
-	}
-	ts := treemap.NewWithIntComparator()
-	ts.Put(nums[0], nil)
-	for j := 1; j < n-1; j++ {
-		if right[j+1] > nums[j] {
-			val, _ := ts.Floor(nums[j] - 1)
-			if val != nil {
-				ans = max(ans, val.(int)-nums[j]+right[j+1])
-			}
-		}
-		ts.Put(nums[j], nil)
-	}
-	return
-}
-```
-
-```ts
 function maximumTripletValue(nums: number[]): number {
     const n = nums.length;
     const right: number[] = Array(n).fill(nums[n - 1]);
@@ -661,8 +508,3 @@ class TreeSet<T = number> {
         return undefined;
     }
 }
-```
-
-<!-- tabs:end -->
-
-<!-- end -->
